refactor: Extract admin entity management logic into dedicated useAdminX hooks and remove generic resource components.
