+++ src/pages/admin/resource/AdminInterestPage.tsx
@@ -1,7 +1,5 @@
-import { useEffect, useState } from 'react';
 import styled from '@emotion/styled';
-import { getInterests, addInterest, deleteInterest } from '../../../api/adminResource';
-import type { Interest } from '../../../types/resource';
+import { useAdminInterests } from '../../../hooks/admin/useAdminInterests';
 import Spinner from '../../../components/common/Spinner';
 
 const Container = styled.div`
@@ -109,50 +107,15 @@ const DeleteButton = styled.button`
 `;
 
 const AdminInterestPage = () => {
-    const [interests, setInterests] = useState<Interest[]>([]);
-    const [newInterest, setNewInterest] = useState('');
-    const [isLoading, setIsLoading] = useState(false);
-    const [isAdding, setIsAdding] = useState(false);
-
-    const fetchInterests = async () => {
-        setIsLoading(true);
-        try {
-            const data = await getInterests();
-            setInterests(data);
-        } catch (error) {
-            console.error('Failed to fetch interests', error);
-        } finally {
-            setIsLoading(false);
-        }
-    };
-
-    useEffect(() => {
-        fetchInterests();
-    }, []);
-
-    const handleAdd = async () => {
-        if (!newInterest.trim()) return;
-        setIsAdding(true);
-        try {
-            await addInterest(newInterest);
-            setNewInterest('');
-            fetchInterests();
-        } catch (error) {
-            alert('관심요소 추가 실패');
-        } finally {
-            setIsAdding(false);
-        }
-    };
-
-    const handleDelete = async (id: number) => {
-        if (!window.confirm('정말로 삭제하시겠습니까?')) return;
-        try {
-            await deleteInterest(id);
-            fetchInterests();
-        } catch (error) {
-            alert('삭제 실패');
-        }
-    };
+    const {
+        interests,
+        newInterest,
+        setNewInterest,
+        isLoading,
+        isAdding,
+        handleAdd,
+        handleDelete
+    } = useAdminInterests();
 
     return (
         <Container>
