+++ src/pages/admin/resource/AdminJobRolePage.tsx
@@ -1,7 +1,5 @@
-import { useEffect, useState } from 'react';
 import styled from '@emotion/styled';
-import { getJobRoles, addJobRole, deleteJobRole } from '../../../api/adminResource';
-import type { JobRole } from '../../../types/resource';
+import { useAdminJobRoles } from '../../../hooks/admin/useAdminJobRoles';
 import Spinner from '../../../components/common/Spinner';
 
 const Container = styled.div`
@@ -109,50 +107,15 @@ const DeleteButton = styled.button`
 `;
 
 const AdminJobRolePage = () => {
-    const [jobRoles, setJobRoles] = useState<JobRole[]>([]);
-    const [newJobRole, setNewJobRole] = useState('');
-    const [isLoading, setIsLoading] = useState(false);
-    const [isAdding, setIsAdding] = useState(false);
-
-    const fetchJobRoles = async () => {
-        setIsLoading(true);
-        try {
-            const data = await getJobRoles();
-            setJobRoles(data);
-        } catch (error) {
-            console.error('Failed to fetch job roles', error);
-        } finally {
-            setIsLoading(false);
-        }
-    };
-
-    useEffect(() => {
-        fetchJobRoles();
-    }, []);
-
-    const handleAdd = async () => {
-        if (!newJobRole.trim()) return;
-        setIsAdding(true);
-        try {
-            await addJobRole(newJobRole);
-            setNewJobRole('');
-            fetchJobRoles();
-        } catch (error) {
-            alert('직무 추가 실패');
-        } finally {
-            setIsAdding(false);
-        }
-    };
-
-    const handleDelete = async (id: number) => {
-        if (!window.confirm('정말로 삭제하시겠습니까?')) return;
-        try {
-            await deleteJobRole(id);
-            fetchJobRoles();
-        } catch (error) {
-            alert('삭제 실패');
-        }
-    };
+    const {
+        jobRoles,
+        newJobRole,
+        setNewJobRole,
+        isLoading,
+        isAdding,
+        handleAdd,
+        handleDelete
+    } = useAdminJobRoles();
 
     return (
         <Container>
